# docs/swagger.yaml
definitions:
...
  dtos.InputOrders:
    properties:
      address:
        type: string
      is_pickup:
        type: boolean
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/dtos.OrderItemCreateInput'
        type: array
      package_id:
        type: integer
      service_id:
        type: integer
      user_id:
        type: integer
      weight:
        description: Optional jika bukan per kg
        type: number
    required:
    - package_id
    - service_id
    type: object
...
  dtos.OrderItemCreateInput:
    properties:
      individual_package_id:
        type: integer
      qty:
        minimum: 1
        type: integer
    required:
    - individual_package_id
    - qty
    type: object
  dtos.OrderItemResponse:
    properties:
      id:
        type: integer
      individual_package_id:
        type: integer
      qty:
        type: integer
      sub_total:
        type: number
    type: object
...
  dtos.ResOrders:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_pickup:
        type: boolean
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/dtos.OrderItemResponse'
        type: array
      package_id:
        type: integer
      service_id:
        type: integer
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
      weight:
        type: number
    type: object
...
  helpers.Pagination:
    properties:
      current_page:
        example: 1
        type: integer
      next_page:
        example: 2
        type: integer
      page_size:
        example: 5
        type: integer
      previous_page:
        example: 0
        type: integer
      total_data:
        example: 50
        type: integer
      total_page:
        example: 10
        type: integer
    type: object
...
  helpers.ResponseCUDSuccess:
    properties:
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
  helpers.ResponseError:
    properties:
      data: {}
      message:
        example: Error message
        type: string
      status:
        example: false
        type: boolean
    type: object
  helpers.ResponseGetAllSuccess:
    properties:
      data: {}
      message:
        example: success message
        type: string
      pagination:
        $ref: '#/definitions/helpers.Pagination'
      status:
        example: true
        type: boolean
    type: object
  helpers.ResponseGetDetailSuccess:
    properties:
      data: {}
      message:
        example: success message
        type: string
      status:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: API Documentation for Cleanease
  title: Example API
  version: "2.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get all orders with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 5
        description: Page size
        in: query
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get all orders success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ResponseGetAllSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtos.ResOrders'
                  type: array
                pagination:
                  $ref: '#/definitions/helpers.Pagination'
              type: object
        "400":
          description: Invalid pagination data
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "404":
          description: No orders found
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ResponseError'
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new laundry order
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dtos.InputOrders'
      produces:
      - application/json
      responses:
        "200":
          description: Create order success
          schema:
            $ref: '#/definitions/helpers.ResponseCUDSuccess'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ResponseError'
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Delete order success
          schema:
            $ref: '#/definitions/helpers.ResponseCUDSuccess'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ResponseError'
      summary: Delete order
      tags:
      - Orders
    get:
      consumes:
      - application/json
      description: Get detailed information of a specific order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get order detail success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ResponseGetDetailSuccess'
            - properties:
                data:
                  $ref: '#/definitions/dtos.ResOrders'
              type: object
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ResponseError'
      summary: Get order details
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Update an existing order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dtos.InputOrders'
      produces:
      - application/json
      responses:
        "200":
          description: Update order success
          schema:
            $ref: '#/definitions/helpers.ResponseCUDSuccess'
        "400":
          description: Invalid request data or order ID
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/helpers.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ResponseError'
      summary: Update order
      tags:
      - Orders
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"